New maps ya se leen de maps.yaml 	- se pueden crear mapas custom y facilmente ponerlos en el server para jugarlos

+++ client/menu_objects/create_match_button.cpp
@@ -14,7 +14,7 @@ void CreateMatchButton::on_click() {
 #ifdef LOG
     std::cout << "Clicked Create match." << std::endl;
 #endif
-    message_handler.create_match(LORI_CHARACTER, (uint8_t) 1, MAX_PLAYERS);
+    message_handler.create_match(LORI_CHARACTER, (uint16_t)1, MAX_PLAYERS);
 }
 
 CreateMatchButton::~CreateMatchButton() = default;
